发布前bug修改

+++ src/Fx.Amiya.Background.Api/Vo/LiveReplayInteractionlData/Input/AddLiveReplayInteractionlDataVo.cs
@@ -22,7 +22,7 @@ namespace Fx.Amiya.Background.Api.Vo.LiveReplayInteractionlData.Input
         /// <summary>
         /// 同比数据
         /// </summary>
-        public int LastLivingData { get; set; }
+        public decimal LastLivingData { get; set; }
         /// <summary>
         /// 同比增长
         /// </summary>
