+++ src/Fx.Amiya.Background.Api/Vo/LiveReplayInteractionlData/Result/LiveLiveReplayInteractionlDataVo.cs
@@ -26,7 +26,7 @@ namespace Fx.Amiya.Background.Api.Vo.LiveReplayInteractionlData.Result
         /// <summary>
         /// 同比数据
         /// </summary>
-        public int LastLivingData { get; set; }
+        public decimal LastLivingData { get; set; }
         /// <summary>
         /// 同比增长
         /// </summary>
