+++ src/Fx.Amiya.Dto/LiveReplayProductDealData/Result/LiveReplayProductDealDataInfoDto.cs
@@ -23,7 +23,7 @@ namespace Fx.Amiya.Dto.LiveReplayProductDealData.Result
         /// <summary>
         /// 同比数据
         /// </summary>
-        public int LastLivingData { get; set; }
+        public decimal LastLivingData { get; set; }
         /// <summary>
         /// 同比增长
         /// </summary>
